Refactor: Organize project structure
- organize project into separate folders -
for tabs and their utility files, main window
and common code
- Rename the most important files and classes to
more descriptive names - to better reflect their
content and purpose
- Add a config file setting up the path to
dependencies
