+++ Tabs/pin_out/Mech_wyj_tuleje/popups.py
